Functions 'varOf', 'attributeOf', etc moved in 'classes.py' (instead of 'driver.py')
