+++ climaf/driver.py
@@ -96,7 +96,7 @@ def maketree(script_name, script, *operands, **parameters):
     rep=classes.ctree(script_name, script, *operands, **parameters)
     # TBD Analyze script inputs cardinality vs actual arguments
     # Create one child for each output
-    defaultVariable=varOf(operands[0])
+    defaultVariable=classes.varOf(operands[0])
         #defaultPeriod=operands[0].period
     for outname in script.outputs :
         if outname is None  or outname=='' :
@@ -247,7 +247,7 @@ def ceval(cobject, userflags=None, format="MaskedArray",
             clogger.info("Object found in cache: %s is at %s:  "%(cobject.crs,filename))
             cdedent()
             if format=='file' : return filename
-            else: return cread(filename,varOf(cobject))
+            else: return cread(filename,classes.varOf(cobject))
         if not isinstance(cobject,classes.cpage) and not isinstance(cobject,classes.cpage_pdf) and \
                not isinstance(cobject,classes.cens) :
             #
@@ -294,7 +294,7 @@ def ceval(cobject, userflags=None, format="MaskedArray",
                 file=ceval_script(cobject,deep,recurse_list=recurse_list) # Does return a filename, or list of filenames
                 cdedent()
                 if ( format == 'file' ) : return (file)
-                else : return cread(file,varOf(cobject))
+                else : return cread(file,classes.varOf(cobject))
             elif cobject.operator in operators.operators :
                 obj=ceval_operator(cobject,deep)
                 cdedent()
@@ -381,13 +381,13 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
             raise Climaf_Driver_Error("Internal error : some input file does not exist among %s:"%(infile))
         subdict[ label ]=infile
         #if scriptCall.flags.canSelectVar :
-        subdict["var"]=varOf(op)
+        subdict["var"]=classes.varOf(op)
         if isinstance(op,classes.cdataset) and op.alias:
             filevar,scale,offset,units,filenameVar,missing=op.alias
-            if scriptCall.flags.canAlias and "," not in varOf(op) :
-                #if script=="select" and ((varOf(op) != filevar) or scale != 1.0 or offset != 0.) :
+            if scriptCall.flags.canAlias and "," not in classes.varOf(op) :
+                #if script=="select" and ((classes.varOf(op) != filevar) or scale != 1.0 or offset != 0.) :
                 subdict["var"]=Template(filevar).safe_substitute(op.kvp)
-                subdict["alias"]="%s,%s,%.4g,%.4g"%(varOf(op),subdict["var"],scale,offset)
+                subdict["alias"]="%s,%s,%.4g,%.4g"%(classes.varOf(op),subdict["var"],scale,offset)
             if units : subdict["units"]=units 
             if scriptCall.flags.canMissing and missing :
                 subdict["missing"]=missing
@@ -404,7 +404,7 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
                 subdict["period"]=str(per)
                 subdict["period_iso"]=per.iso()
         if scriptCall.flags.canSelectDomain :
-            subdict["domain"]=domainOf(op)
+            subdict["domain"]=classes.domainOf(op)
     i=0
     for op in scriptCall.operands :
         if op : opscrs += op.crs+" - "
@@ -419,13 +419,13 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
             label,multiple,serie=script.inputs[i]
             subdict[ label ]=infile
             # Provide the name of the variable in input file if script allows for
-            subdict["var_%d"%i]=varOf(op)
+            subdict["var_%d"%i]=classes.varOf(op)
             if isinstance(op,classes.cdataset) and op.alias :
                 filevar,scale,offset,units,filenameVar,missing =op.alias
-                if ((varOf(op) != filevar) or (scale != 1.0) or (offset != 0.)) and \
-                        "," not in varOf(op):
+                if ((classes.varOf(op) != filevar) or (scale != 1.0) or (offset != 0.)) and \
+                        "," not in classes.varOf(op):
                     subdict["var_%d"%i]=Template(filevar).safe_substitute(op.kvp)
-                    subdict["alias_%d"%i]="%s %s %f %f"%(varOf(op),subdict["var_%d"%i],scale,offset)
+                    subdict["alias_%d"%i]="%s %s %f %f"%(classes.varOf(op),subdict["var_%d"%i],scale,offset)
                 if units : subdict["units_%d"%i]=units 
                 if missing : subdict["missing_%d"%i]=missing
                 # Provide period selection if script allows for
@@ -434,7 +434,7 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
                 if not per.fx and per != "":
                     subdict["period_%d"%i]=str(per)
                     subdict["period_iso_%d"%i]=per.iso()
-            subdict["domain_%d"%i]=domainOf(op)
+            subdict["domain_%d"%i]=classes.domainOf(op)
     clogger.debug("subdict for operands is "+`subdict`)
     # substitution is deffered after scriptcall parameters evaluation, which may
     # redefine e.g period
@@ -456,10 +456,10 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
         main_output_filename=tempfile.NamedTemporaryFile(suffix="."+output_fmt).name #cache.generateUniqueFileName(scriptCall.crs, format=output_fmt)
   
         subdict["out"]=main_output_filename
-        subdict["out_"+varOf(scriptCall)]=main_output_filename
+        subdict["out_"+classes.varOf(scriptCall)]=main_output_filename
 
         subdict["out_final"]=cache.generateUniqueFileName(scriptCall.crs, format=output_fmt) 
-        subdict["out_final_"+varOf(scriptCall)]=cache.generateUniqueFileName(scriptCall.crs, format=output_fmt) 
+        subdict["out_final_"+classes.varOf(scriptCall)]=cache.generateUniqueFileName(scriptCall.crs, format=output_fmt) 
         
         # Named outputs
         for output in scriptCall.outputs:
@@ -504,11 +504,11 @@ def ceval_script (scriptCall,deep,recurse_list=[]):
     if script.fixedfields is not None :
         #subdict_ff=dict()
         subdict_ff=scriptCall.parameters.copy()
-        subdict_ff["model"]=modelOf(scriptCall.operands[0])
-        subdict_ff["simulation"]=simulationOf(scriptCall.operands[0])
-        subdict_ff["project"]=projectOf(scriptCall.operands[0])
-        subdict_ff["realm"]=realmOf(scriptCall.operands[0])
-        subdict_ff["grid"]=gridOf(scriptCall.operands[0])
+        subdict_ff["model"]=classes.modelOf(scriptCall.operands[0])
+        subdict_ff["simulation"]=classes.simulationOf(scriptCall.operands[0])
+        subdict_ff["project"]=classes.projectOf(scriptCall.operands[0])
+        subdict_ff["realm"]=classes.realmOf(scriptCall.operands[0])
+        subdict_ff["grid"]=classes.gridOf(scriptCall.operands[0])
         l=script.fixedfields #return paths: (linkname, targetname)
         files_exist=dict()
         for ll,lt in l:
@@ -572,59 +572,7 @@ def timePeriod(cobject) :
         clogger.debug("for now, timePeriod logic for 'cens' objet is basic (1st member)- TBD")
         return timePeriod(cobject.values()[0])
     else : return None #clogger.error("unkown class for argument "+`cobject`)
-                  
-def domainOf(cobject) :
-    """ Returns a domain for a CliMAF object : if object is a dataset, returns
-    its domain, otherwise returns domain of first operand
-    """
-    if isinstance(cobject,classes.cdataset) : 
-	if type(cobject.domain) is list :
-            rep=""
-            for coord in cobject.domain[0:-1] : rep=r"%s%d,"%(rep,coord)
-            rep="%s%d"%(rep,cobject.domain[-1])
-            return(rep)
-	else : 
-	    if cobject.domain == "global" : return ""
-	    else : return(cobject.domain)
-    elif isinstance(cobject,classes.ctree) :
-        clogger.debug("For now, domainOf logic for scripts output is basic (1st operand) - TBD")
-        return domainOf(cobject.operands[0])
-    elif isinstance(cobject,classes.scriptChild) :
-        clogger.debug("For now, domainOf logic for scriptChilds is basic - TBD")
-        return domainOf(cobject.father)
-    elif isinstance(cobject,classes.cens) :
-        clogger.debug("for now, domainOf logic for 'cens' objet is basic (1st member)- TBD")
-        return domainOf(cobject.values()[0])
-    elif cobject is None : return "none"
-    else : clogger.error("Unkown class for argument "+`cobject`)
-                  
-
-def varOf(cobject) : return attributeOf(cobject,"variable")
-def modelOf(cobject) : return attributeOf(cobject,"model")
-def simulationOf(cobject) : return attributeOf(cobject,"simulation")
-def projectOf(cobject) : return attributeOf(cobject,"project")
-def realmOf(cobject) : return attributeOf(cobject,"realm")
-def gridOf(cobject) : return attributeOf(cobject,"grid")
-
-def attributeOf(cobject,attrib) :
-    """ Returns the attribute for a CliMAF object : if object is a dataset, returns
-    its attribute property, otherwise returns attribute of first operand
-    """
-    if isinstance(cobject,classes.cdataset) : 
-        val=getattr(cobject,attrib,None) 
-        if val is not None : return val
-        else : return(cobject.kvp.get(attrib))
-    elif isinstance(cobject,classes.cens) : return attributeOf(cobject.values()[0],attrib)
-    elif getattr(cobject,attrib,None) : return getattr(cobject,attrib) 
-    elif isinstance(cobject,classes.ctree) :
-        clogger.debug("for now, varOf logic is basic (1st operand) - TBD")
-        return attributeOf(cobject.operands[0],attrib)
-    elif isinstance(cobject,cmacro.cdummy) :
-        return "dummy"
-    elif isinstance(cobject,classes.cpage) or isinstance(cobject,classes.cpage_pdf) : return None
-    elif cobject is None : return ''
-    else : raise Climaf_Driver_Error("Unknown class for argument "+`cobject`)
-
+ 
                   
 def ceval_select(includer,included,userflags,format,deep,derived_list,recurse_list) :
     """ Extract object INCLUDED from (existing) object INCLUDER,
@@ -1124,7 +1072,7 @@ def efile(obj, filename, forced=False) :
             ffile=cfile(memb)
             
             f = tempfile.NamedTemporaryFile(suffix=".nc")
-            command="ncrename -O -v %s,%s_%s %s %s"%(varOf(memb), varOf(memb), lab, ffile, f.name)
+            command="ncrename -O -v %s,%s_%s %s %s"%(classes.varOf(memb), classes.varOf(memb), lab, ffile, f.name)
             if ( os.system(command) != 0 ) :
                 raise Climaf_Driver_Error("ncrename failed : %s" %command)
